aliased 'unsigned long long' to 'ull'

+++ beowolf/math/W_Math.h
@@ -5,7 +5,7 @@ namespace wolf
 	class Math
 	{
 	public:
-		static bool isPrime(unsigned long long num);
+		static bool isPrime(ull num);
 		static float lerp(float start, float end, float percent);
 		static double Math::lerp(double start, double end, float percent);
 	};
